fix(paste) helper function to dispatch paste to replace actions
