+++ editor/src/components/canvas/canvas-strategies/post-action-options/post-action-options.ts
@@ -1,9 +1,17 @@
 import type { BuiltInDependencies } from '../../../../core/es-modules/package-manager/built-in-dependencies-list'
 import { stripNulls } from '../../../../core/shared/array-utils'
+import type { ElementPath } from '../../../../core/shared/project-file-types'
 import { assertNever } from '../../../../core/shared/utils'
+import type { EditorAction } from '../../../editor/action-types'
+import {
+  executePostActionMenuChoice,
+  startPostActionSession,
+} from '../../../editor/actions/action-creators'
 import type {
   DerivedState,
   EditorState,
+  InternalClipboard,
+  PasteToReplacePostActionMenuData,
   PostActionMenuData,
 } from '../../../editor/store/editor-state'
 import type { CanvasCommand } from '../../commands/commands'
@@ -46,4 +54,28 @@ export function generatePostactionChoices(data: PostActionMenuData): PostActionC
     default:
       assertNever(data)
   }
+}
+
+export function createPasteToReplacePostActionActions(
+  selectedViews: Array<ElementPath>,
+  internalClipboard: InternalClipboard,
+): Array<EditorAction> | null {
+  const pasteToReplacePostActionMenuData: PasteToReplacePostActionMenuData = {
+    type: 'PASTE_TO_REPLACE',
+    targets: selectedViews,
+    internalClipboard: internalClipboard,
+  }
+
+  const defaultChoice = stripNulls([
+    PasteToReplaceWithPropsReplacedPostActionChoice(pasteToReplacePostActionMenuData),
+    PasteToReplaceWithPropsPreservedPostActionChoice(pasteToReplacePostActionMenuData),
+  ]).at(0)
+
+  if (defaultChoice != null) {
+    return [
+      startPostActionSession(pasteToReplacePostActionMenuData),
+      executePostActionMenuChoice(defaultChoice),
+    ]
+  }
+  return null
 }
